Create directory Plugin/CompilerLocator and move associated code inside it.
