+++ Plugin/CompilerLocator/CompilerLocatorCLANG.cpp
