+++ Plugin/CompilerLocator/CompilerLocatorCLANG.h
