+++ Plugin/CompilerLocator/CompilerLocatorCrossGCC.cpp
