+++ Plugin/CompilerLocator/CompilerLocatorCygwin.cpp
