+++ Plugin/CompilerLocator/CompilerLocatorCygwin.h
