+++ Plugin/CompilerLocator/CompilerLocatorEosCDT.cpp
