+++ Plugin/CompilerLocator/CompilerLocatorGCC.cpp
