+++ Plugin/CompilerLocator/CompilerLocatorMSVC.cpp
