+++ Plugin/CompilerLocator/CompilerLocatorMSVC.h
