+++ Plugin/CompilerLocator/CompilerLocatorMSYS2.cpp
