+++ Plugin/CompilerLocator/CompilerLocatorMSYS2.hpp
