+++ Plugin/CompilerLocator/CompilerLocatorMSYS2Clang.cpp
