+++ Plugin/CompilerLocator/CompilerLocatorMSYS2Clang.hpp
