+++ Plugin/CompilersDetectorManager.cpp
@@ -25,15 +25,15 @@
 
 #include "CompilersDetectorManager.h"
 
-#include "CompilerLocatorCLANG.h"
-#include "CompilerLocatorCrossGCC.h"
-#include "CompilerLocatorCygwin.h"
-#include "CompilerLocatorEosCDT.h"
-#include "CompilerLocatorGCC.h"
-#include "CompilerLocatorMSVC.h"
-#include "CompilerLocatorMSYS2.hpp"
-#include "CompilerLocatorMSYS2Clang.hpp"
-#include "CompilerLocatorRustc.hpp"
+#include "CompilerLocator/CompilerLocatorCLANG.h"
+#include "CompilerLocator/CompilerLocatorCrossGCC.h"
+#include "CompilerLocator/CompilerLocatorCygwin.h"
+#include "CompilerLocator/CompilerLocatorEosCDT.h"
+#include "CompilerLocator/CompilerLocatorGCC.h"
+#include "CompilerLocator/CompilerLocatorMSVC.h"
+#include "CompilerLocator/CompilerLocatorMSYS2.hpp"
+#include "CompilerLocator/CompilerLocatorMSYS2Clang.hpp"
+#include "CompilerLocator/CompilerLocatorRustc.hpp"
 #include "GCCMetadata.hpp"
 #include "JSON.h"
 #include "build_settings_config.h"
